requirements.txt
streamlit>=1.35
pillow>=10.0
numpy>=1.24
opencv-python>=4.8
pymupdf>=1.23
python-docx>=1.1
paddlepaddle
paddleocr
vietocr
torch
torchvision
transformers>=4.40
huggingface-hub>=0.34.0,<1.0
sentencepiece
jiwer
